+++ xcube/api/verify.py
@@ -24,6 +24,15 @@ def verify_cube(dataset: xr.Dataset) -> List[str]:
     """
     Verify the given *dataset* for being a valid data cube.
 
+    The tool verifies that *dataset*
+    * defines the dimensions "time", "lat", "lon";
+    * has corresponding "time", "lat", "lon" coordinate variables and that they
+      are valid, e.g. 1-D, non-empty, using correct units;
+    * has valid  bounds variables for "time", "lat", "lon" coordinate
+      variables, if any;
+    * has any data variables and that they are valid, e.g. min. 3-D, all have
+      same dimensions, have at least dimensions "time", "lat", "lon".
+
     Returns a list of issues, which is empty if *dataset* is a valid data cube.
 
     :param dataset: A dataset to be verified.